MOONCAKE: move the TENT request path into its own translation unit
postXferTent, checkXferTent and releaseReqHTent move to
mooncake_backend_tent.cpp verbatim, with the request handle and the batch
constants both halves need going to a new mooncake_backend_internal.h.
mooncake_backend.cpp keeps the classic path and the mode dispatch, which is
what it was getting hard to read next to the TENT bodies.

Then the review comments:

releaseReqH() refused the release with NIXL_ERR_REPOST_ACTIVE while cancelled
transfers were still in flight, expecting the caller to come back. It does,
but not through the same path: nixlAgent::releaseXferReq() stores the refusal
in the request status, so the next call no longer sees NIXL_IN_PROG, skips the
release block and deletes the handle, and ~nixlXferReqH() discards what
releaseReqH() returns. A second refusal therefore stranded the batch in the
engine with its transfers still running. It now refuses exactly once: the
second release waits up to 100 ms for the cancellation to reach a terminal
state and frees the batch either way, logging when it had to. Leaking the
batch is the worse of the two outcomes, and that call is the core's last.

deregisterMem's else branch carries braces, matching registerMem above it and
the brace rule in docs/CodeStyle.md.

Mode/Classic/Tent are renamed to mode/CLASSIC/TENT for the EnumCase and
EnumConstantCase rules in .clang-tidy.

A static plugin carries no link_args of its own, so --no-as-needed never
reached the libnixl link line in a static-plugin build; it now rides the
exported dependency. Meson emits dependency link_args ahead of the libraries
they apply to, so it lands in front of -ltent_shared as it has to.
